Adding missing pointer casts
